Update of sorting values

+++ src/pbrt/estimators.h
@@ -82,8 +82,8 @@ struct PixelWindow {
 
     // use for sort values
     int indices[maxnbuffers];
-    Float sortedValues[maxnbuffers];
-    Float cvalues[maxnbuffers];
+    double sortedValues[maxnbuffers];
+    double cvalues[maxnbuffers];
     Float weightsSum[maxnbuffers];
     double csplats[maxnbuffers];
